+++ packages/hardhat/deploy-helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './addresses';
+export * from './helpers';
