+++ packages/hardhat/deploy/001_diamond.ts
@@ -1,6 +1,6 @@
 import { DeployFunction } from 'hardhat-deploy/types';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { deploy, verify } from './helpers';
+import { deploy, verify } from '../deploy-helpers';
 
 type SifiDiamondDeployArgs = [owner: string, diamondCutFacetAddress: string];
